refactor collection folder tree
