+++ packages/veritone-react-common/src/components/SimpleSearchBar/styles.js
@@ -32,4 +32,7 @@ export default {
     justifyContent: 'flex-end',
     alignItems: 'flex-end',
   },
+  focusedColor: {
+    backgroundColor: '#fff'
+  }
 }
