+++ packages/veritone-widgets/src/redux/modules/folder/sagas/createFolderSagas.js
@@ -5,15 +5,15 @@ import {
   select
 } from 'redux-saga/effects';
 import get from 'lodash/get';
-import { handleRequest } from './helper';
-import * as folderReducer from './index';
-import * as folderSelector from './selector';
+import { handleRequest } from '../helper';
+import * as actions from '../actions';
+import * as folderSelector from '../selector';
 export default function* createFolderSaga() {
-  yield takeEvery(folderReducer.CREATE_FOLDER, createFolder);
+  yield takeEvery(actions.CREATE_FOLDER, createFolder);
 }
 
 function* createFolder(action) {
-  yield put(folderReducer.createFolderStart());
+  yield put(actions.createFolderStart());
   const {
     name,
     parentId
@@ -39,10 +39,10 @@ function* createFolder(action) {
   }
   const { error, response } = yield call(handleRequest, { query, variables });
   if (error) {
-    yield put(folderReducer.createFolderError());
+    yield put(actions.createFolderError());
   }
   const newFolderId = get(response, 'data.createFolder.id');
-  yield put(folderReducer.initFolder(newFolderId));
-  yield put(folderReducer.initFolder(parentId));
-  return put(folderReducer.createFolderSuccess());
+  yield put(actions.initFolder(newFolderId));
+  yield put(actions.initFolder(parentId));
+  return put(actions.createFolderSuccess());
 }
