+++ packages/veritone-widgets/src/widgets/AppBar/index.js
@@ -61,6 +61,7 @@ class AppBar extends React.Component {
   };
 
   render() {
+    console.log(this.props);
     return (
       <Fragment>
         <LibAppBar
